VCFTools_BCFTools_GBS.sh
VCFTools is used to filter VCFs to be biallelic, have a mean-minimum depth of 5x, and maximum missingness of 0.9.
BCFTools is used to prune based on LD r2 of 0.8 and a sliding window of 1000bps. 
VCFTools is used to thin VCFs so there is a minimum of 64bp between SNPs.